pyproject.toml
[project]
name = "bot-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
